Make createServer an express app factory
Move to src/server.js
